Fix error with running setup.sh twice and add more logging

+++ setup.sh
@@ -112,8 +112,8 @@ if [ -f "src/JBrowse/Browser.js" ]; then
     (
         set -e
         check_node
-        npm install yarn
-        node_modules/.bin/yarn install
+        [[ -f node_modules/.bin/yarn ]] || npm install yarn --no-progress --verbose
+        node_modules/.bin/yarn install --no-progress --verbose
         node_modules/.bin/yarn build
     ) >>setup.log 2>&1;
     done_message "" "" "FAILURE NOT ALLOWED"
